[BUG] Handle when no pricing information available
Sometimes beer is not listed with a price, e.g. instead of a price the text "Ask" is shown.

This was causing an error as it was assumed it would always be a number.

This has now been changed to handle that scenario by reporting the beer price as `null`, and the formatted price as `'Unknown'`.