[BUG] Fix price parsing issue when no pence value
When a drink price is specified as an integer value, e.g. "£5", it was not being parsed correctly as the library previously assumed that a pence value would be specified.

As a result, the price would be reported as "null", and the formatted price reported as "Unknown".

This fixes the issue and adds a test case for it.

Given a price of "£5" on the beer board, the drink price will now be reported as "5" and the formatted price reported as "£5.00".